+++ packages/expo-sharing/ios/SharingModule.swift
@@ -6,14 +6,9 @@ public final class SharingModule: Module {
     Name("ExpoSharing")
 
     AsyncFunction("shareAsync") { (url: URL, options: SharingOptions, promise: Promise) in
-      guard let filePermissions: EXFilePermissionModuleInterface =
-        appContext?.legacyModule(implementing: EXFilePermissionModuleInterface.self)
-      else {
-        throw FilePermissionModuleException()
-      }
+      let grantedPermissions = FileSystemUtilities.permissions(appContext, for: url)
 
-      let grantedPermissions = filePermissions.getPathPermissions(url.relativePath)
-      guard grantedPermissions.rawValue >= EXFileSystemPermissionFlags.read.rawValue else {
+      guard grantedPermissions.contains(.read) else {
         throw FilePermissionException()
       }