+++ packages/expo-video-thumbnails/ios/VideoThumbnailsModule.swift
@@ -12,11 +12,7 @@ public class VideoThumbnailsModule: Module {
 
   internal func getVideoThumbnail(sourceFilename: URL, options: VideoThumbnailsOptions) throws -> [String: Any] {
     if sourceFilename.isFileURL {
-      guard let fileSystem = self.appContext?.fileSystem else {
-        throw Exceptions.FileSystemModuleNotFound()
-      }
-
-      guard fileSystem.permissions(forURI: sourceFilename).contains(.read) else {
+      guard FileSystemUtilities.permissions(appContext, for: sourceFilename).contains(.read) else {
         throw FileSystemReadPermissionException(sourceFilename.absoluteString)
       }
     }
@@ -44,18 +40,18 @@ public class VideoThumbnailsModule: Module {
   Saves the image as a file.
   */
   internal func saveImage(image: UIImage, quality: Double) throws -> URL {
-    guard let fileSystem = self.appContext?.fileSystem else {
-      throw Exceptions.FileSystemModuleNotFound()
-    }
-
-    let directory = URL(fileURLWithPath: fileSystem.cachesDirectory).appendingPathComponent("VideoThumbnails")
+    let directory = appContext?.config.cacheDirectory?.appendingPathComponent("VideoThumbnails")
     let fileName = UUID().uuidString.appending(".jpg")
-    let fileUrl = directory.appendingPathComponent(fileName)
+    let fileUrl = directory?.appendingPathComponent(fileName)
 
-    fileSystem.ensureDirExists(withPath: directory.path)
+    FileSystemUtilities.ensureDirExists(at: directory)
 
     guard let data = image.jpegData(compressionQuality: CGFloat(quality)) else {
       throw CorruptedImageDataException()
+    }
+
+    guard let fileUrl else {
+      throw ImageWriteFailedException("Unrecognized url \(fileUrl?.path)")
     }
 
     do {